Support sequential reads from PASSIVE members.
